Level editor: voxel terrain edit seam and planet sculpting (Milestone 2)
Gives `TerrainMode` a real sculpt pipeline against the planetary voxel
runtime. Raise/Lower/Flatten/Paint now mutate planet voxels through a new
`TerrainEditApi` seam, with a brush ring drawn in the viewport, per-stroke
undo, and terrain that survives a save/load.

engine_backend::services::terrain_edit (new)
  The editor's only door into `pulsar_terrain`: `hit_terrain(ray)`,
  `apply_edit(target, op)`, snapshot/restore, and edit-log import/export.
  Every terrain type a UI crate needs is re-exported from here, so no UI
  crate takes its own `pulsar_terrain` dependency and no pointer handler
  ever sees a raw `TerrainRuntimeHandle`.

  Threading follows the existing conventions, with a deliberate split:
  mutations go straight through `TerrainRuntimeHandle` (its own designed
  cross-thread surface -- and the caller needs the allocated mutation
  sequence synchronously to answer Consumed vs PassThrough), while
  anything touching the *render thread* goes through `TerrainEditMailbox`,
  shaped exactly like `HelioEditorMailbox`. No new ad-hoc locking.

  `apply_edit` rewrites each op's sequence and stable id: the runtime
  rejects any operation that does not strictly advance the planet's
  sequence, and only the runtime knows the current value.

Planet activation
  `HelioInner::planet_terrain` has been permanently `None` since ddad02d7
  removed the generic world-component dispatch, and `sync_planet_graph`
  had no callers -- so there is no planet in the editor at all today.
  Rather than block this milestone on that being restored, the seam
  re-supplies it narrowly: the editor collects `PlanetTerrainComponent`s
  from its own scene database and posts them through the mailbox, and the
  render thread creates the runtime and upserts them. The source-key
  format matches `PlanetTerrainComponent`'s own `sync_component`, so the
  two paths address the same planet when the generic dispatch returns.

Sculpting
  `TerrainMode::on_pointer` handles Down/Drag/Up/Hover. `Consumed` only
  when the brush had terrain under it, so empty-space clicks still select
  objects. Stamps coalesce below a quarter of the brush radius. Flatten
  anchors to the altitude captured at stroke start. `pulsar_terrain` has
  one brush primitive, so each mode is expressed as sphere placement:
  Raise sinks it to protrude by `strength`, Lower lifts it to bite in,
  Flatten anchors its top to the stroke altitude, Paint centres on the hit.

Undo
  A new bounded `TerrainUndoDomain`, separate from `SceneDomain`'s stack
  because that one snapshots the scene database, which holds no voxels.
  One stroke is one entry: a snapshot anchor plus the stamps it committed
  (append-only log, so snapshot + ops determines the result and redo is
  near-free). Ctrl+Z/Ctrl+Y route to it only while Terrain mode is active.

Persistence
  `save_to_file_with_editor_camera` gains a terrain parameter and flushes
  each planet's mutation log to a `<level>.terrain` sidecar. A sidecar
  rather than a `LevelFile` section: levels without terrain stay
  byte-identical, and bulk binary does not belong in a reviewed JSON
  document. Replay is retried per frame until the render thread has
  registered the planets.

pulsar_terrain gains read-only accessors an external edit producer needs
(`planet_definitions`, `latest_sequence`, `sample_cells`,
`planet_snapshot`) plus `restore_planet_snapshot`, which rewinds the
mutation tail through the existing planet-wide retirement path.

New dependencies: `pulsar_terrain` for engine_backend (the seam re-exports
its types) and `base64` for ui_level_editor (encodes the binary edit log
into the JSON sidecar). Both already workspace dependencies.

`LevelEditMode` is untouched and still always returns `PassThrough`.

Deferred: `create_planet` returns `CreationUnsupported` -- a planet exists
because a scene object carries a `PlanetTerrainComponent`, and authoring
that component is scene work, not terrain editing.

Verified: cargo check --workspace; cargo test -p engine_backend (90
passed), -p pulsar_terrain (9), -p ui_level_editor (82 + 6 + 6). The two
pre-existing toggle_button.rs doctest failures are unrelated.

Closes #711, part of #709.

+++ crates/editor/ui_level_editor/src/level_editor/state/terrain.rs
@@ -81,7 +81,12 @@ pub enum TerrainTarget {
 #[derive(Clone, Copy, Debug, PartialEq, Serialize, Deserialize)]
 pub struct Stroke {
     pub id: u64,
-    pub start_height: f32,
+    /// Altitude the stroke was anchored to, in meters from the planet centre.
+    ///
+    /// `f64`, not `f32`: on an Earth-sized planet this is ~6.4e6, where an
+    /// `f32` step is around half a metre -- enough to make a Flatten stroke
+    /// visibly stair-step instead of levelling.
+    pub start_height: f64,
 }
 
 // ── Terrain Domain ─────────────────────────────────────────────────────────
@@ -120,7 +125,7 @@ impl TerrainDomain {
         self.target = target;
     }
 
-    pub fn begin_stroke(&mut self, id: u64, start_height: f32) {
+    pub fn begin_stroke(&mut self, id: u64, start_height: f64) {
         self.active_stroke = Some(Stroke { id, start_height });
     }
